wallet: fix if statement to also check err type and lightning code

+++ wallet/wallet.go
@@ -913,18 +913,16 @@ func (w *Wallet) Melt(quoteId string) (*nut05.PostMeltQuoteBolt11Response, error
 	}
 	meltBolt11Response, err := client.PostMeltBolt11(mint.mintURL, meltBolt11Request)
 	if err != nil {
-		if cashuErr, ok := err.(cashu.Error); ok {
-			if cashuErr.Code == cashu.LightningPaymentErrCode {
-				// only remove proofs from pending and save them for use
-				// if got specific error that payment failed
-				if err := w.db.SaveProofs(proofs); err != nil {
-					return nil, fmt.Errorf("error storing proofs: %v", err)
-				}
-				if err := w.db.DeletePendingProofsByQuoteId(quote.QuoteId); err != nil {
-					return nil, fmt.Errorf("error removing pending proofs: %v", err)
-				}
-				return nil, err
+		if cashuErr, ok := err.(cashu.Error); ok && cashuErr.Code == cashu.LightningPaymentErrCode {
+			// only remove proofs from pending and save them for use
+			// if got specific error that payment failed
+			if err := w.db.SaveProofs(proofs); err != nil {
+				return nil, fmt.Errorf("error storing proofs: %v", err)
 			}
+			if err := w.db.DeletePendingProofsByQuoteId(quote.QuoteId); err != nil {
+				return nil, fmt.Errorf("error removing pending proofs: %v", err)
+			}
+			return nil, err
 		} else {
 			// for any other errors leave proofs as pending
 			return nil, fmt.Errorf("error doing melt request: %v. Proofs are pending", err)
